added jpa, final version (Server) for Winter Semester

+++ src/main/java/repository/Repository.java
@@ -10,11 +10,15 @@ import utils.CustomException;
 import utils.Role;
 import utils.User;
 
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
+
 
 public class Repository {
 
-    //private EntityManagerFactory emf = Persistence.createEntityManagerFactory("studentsVotePU");
-    //private EntityManager em = emf.createEntityManager();
+    private EntityManagerFactory emf = Persistence.createEntityManagerFactory("studentsVotePU");
+    private EntityManager em = emf.createEntityManager();
 
 
 
@@ -36,16 +40,13 @@ public class Repository {
         try {
             LdapUser lu = new LdapUser(user.getUsername(), user.getPassword().toCharArray());
 
-            return  jsonLoginBuilder(lu.getUserId(), Role.Students);
-
         } catch (LdapException e) {
              return ce.buildException(503, "Service Unavailable", "LDAP not working");
-
         } catch (LdapAuthException e) {
             return ce.buildException(401, "Unauthorized", "Login Error");
         }
 
-         /* if(user.getUsername().equals("test") && user.getPassword().equals("1234")) {
+
             if(isCandidate(user.getUsername())){
                 return jsonLoginBuilder(user.getUsername(), Role.Candidates);
             } else if(isReturningOfficer(user.getUsername())){
@@ -53,9 +54,6 @@ public class Repository {
             }
 
             return jsonLoginBuilder(user.getUsername(), Role.Students);
-        }
-
-        */
     }
 
     private String jsonLoginBuilder(String username, Role role) {
@@ -75,32 +73,35 @@ public class Repository {
 
 
     public boolean isCandidate(String username) {
-        //return em.find(Candidate.class, username).getUsername() == null;
-        return false;
+        return em.find(Candidate.class, username).getUsername() == null;
     }
 
     public boolean isReturningOfficer(String username) {
-        //return em.find(ReturningOfficer.class, username).getUsername() == null;
-        return false;
+        return em.find(ReturningOfficer.class, username).getUsername() == null;
     }
 
 
     public String setCandidate(Candidate candidate) {
-        //em.getTransaction().begin();
-        //em.persist(candidate);
-        //em.getTransaction().commit();
+        em.getTransaction().begin();
+        em.persist(candidate);
+        em.getTransaction().commit();
         return "got it";
     }
 
 
-    public String changereturningofficer(ReturningOfficer rsold, ReturningOfficer rsnew) {
-        //em.getTransaction().begin();
-        /*if(em.find(ReturningOfficer.class, rsold) == null){
+    public String changereturningofficer(String username_old, String password_old, String username_new, String password_new) {
+        ReturningOfficer rsold = new ReturningOfficer(1, password_old, username_old);
+        ReturningOfficer rsnew = new ReturningOfficer(1, password_new, username_new);
+
+        em.getTransaction().begin();
+        if(em.find(ReturningOfficer.class, 1) == rsold){
+
             em.remove(rsold);
             em.persist(rsnew);
+            em.getTransaction().commit();
             return "changed Returning Officer";
-        }*/
-
+        }
+        em.getTransaction().commit();
         return "wrong Returning Officer";
     }
 }
